backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
google-cloud-storage==2.13.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
email-validator==2.1.0.post1
